refactor: rename schema package to api
The schema/ package was the user-facing input surface but its name
implied it was about internal schema concerns. Renaming to api/ makes
the role immediately legible alongside the other layers (domain,
application, adapters).
