+++ src/delta_engine/api/__init__.py
@@ -12,8 +12,8 @@ from delta_engine.domain.model import (
     String,
     Timestamp,
 )
-from delta_engine.schema.properties import Property
-from delta_engine.schema.table import DeltaTable
+from delta_engine.api.properties import Property
+from delta_engine.api.table import DeltaTable
 
 __all__ = [
     "Array",
