+++ src/delta_engine/api/properties.py
