+++ tests/application/test_registry.py
@@ -3,7 +3,7 @@ import pytest
 
 from delta_engine.application.registry import Registry
 from delta_engine.domain.model import Column, DesiredTable, Integer, QualifiedName
-from delta_engine.schema import DeltaTable, String
+from delta_engine.api import DeltaTable, String
 
 
 @st.composite
