+++ tests/e2e/test_engine_e2e.py
@@ -8,7 +8,7 @@ from delta_engine.adapters.databricks.reader import DatabricksReader
 from delta_engine.application.engine import Engine
 from delta_engine.application.errors import SyncFailedError
 from delta_engine.application.registry import Registry
-from delta_engine.schema import Column, Date, DeltaTable, Integer, String
+from delta_engine.api import Column, Date, DeltaTable, Integer, String
 from tests.config import TEST_CATALOG
